+++ src/core/adapters/mobileWebAuth/mobileWebAuth.localnet.test.ts
@@ -7,12 +7,13 @@ import { Identity, type MobileWebAuth } from 'wata'
 import { Wata as HostWata, mobileWebAuth as hostMobileWebAuth } from 'wata/host'
 import * as z from 'zod/mini'
 
-import { accounts, chain, getClient } from '../../../test/config.js'
-import * as Provider from '../Provider.js'
-import * as Storage from '../Storage.js'
-import * as Store from '../Store.js'
-import * as Rpc from '../zod/rpc.js'
-import { mobileWebAuth, tempoWallet } from './mobileWebAuth.js'
+import { accounts, chain, getClient } from '../../../../test/config.js'
+import * as Provider from '../../Provider.js'
+import * as Storage from '../../Storage.js'
+import * as Store from '../../Store.js'
+import * as Rpc from '../../zod/rpc.js'
+import { mobileWebAuth } from './mobileWebAuth.js'
+import { tempoWallet } from './tempoWallet.js'
 
 const callback = 'xyz.tempo.accounts.playground:/auth'
 const consumerOrigin = 'https://accounts-playground.example'
