+++ src/index.ts
@@ -15,6 +15,8 @@ export * as TrustedHosts from './core/TrustedHosts.js'
 export * as WebAuthnCeremony from './core/WebAuthnCeremony.js'
 export { dialog, dialog as tempoWallet } from './core/adapters/dialog.js'
 export { local } from './core/adapters/local.js'
+export { mobileWebAuth } from './core/adapters/mobileWebAuth/mobileWebAuth.js'
+export { tempoWallet as tempoWalletMobileWebAuth } from './core/adapters/mobileWebAuth/tempoWallet.js'
 export { turnkey } from './core/adapters/turnkey.js'
 export { webAuthn } from './core/adapters/webAuthn.js'
 export {
