--- src/mobileWebAuth/index.ts
@@ -1,1 +1,0 @@
-export { mobileWebAuth, tempoWallet } from '../core/adapters/mobileWebAuth.js'
