--- src/mobileWebAuth/tempoWallet.ts
@@ -1,1 +1,0 @@
-export { tempoWallet } from '../core/adapters/mobileWebAuth.js'
